+++ payment/src/main/java/com/carrot/payment/controller/PaymentTossApiController.java
@@ -1,0 +1,106 @@
+package com.carrot.payment.controller;
+
+import java.io.*;
+import java.net.HttpURLConnection;
+import java.net.URL;
+import java.nio.charset.StandardCharsets;
+import java.util.Base64;
+
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+
+import org.json.simple.JSONObject;
+import org.json.simple.parser.JSONParser;
+import org.json.simple.parser.ParseException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.ResponseEntity;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequiredArgsConstructor
+public class PaymentTossApiController {
+    private final Logger logger = LoggerFactory.getLogger(this.getClass());
+    private static final String WIDGET_SECRET_KEY = "test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6";
+    private static final String API_SECRET_KEY = "test_sk_mBZ1gQ4YVXjawyl0aAq68l2KPoqN";
+
+    @RequestMapping(value = {"/confirm/widget", "/confirm/payment"})
+    public ResponseEntity<JSONObject> confirmPayment(HttpServletRequest request, @RequestBody String jsonBody) throws Exception {
+        String secretKey = request.getRequestURI().contains("/confirm/payment") ? API_SECRET_KEY : WIDGET_SECRET_KEY;
+        JSONObject response = sendRequest(parseRequestData(jsonBody), secretKey, "https://api.tosspayments.com/v1/payments/confirm");
+        int statusCode = response.containsKey("error") ? 400 : 200;
+        return ResponseEntity.status(statusCode).body(response);
+    }
+
+    @RequestMapping(value = "/callback-auth", method = RequestMethod.GET)
+    public ResponseEntity<JSONObject> callbackAuth(@RequestParam String customerKey, @RequestParam String code) throws Exception {
+        JSONObject requestData = new JSONObject();
+        requestData.put("grantType", "AuthorizationCode");
+        requestData.put("customerKey", customerKey);
+        requestData.put("code", code);
+
+        String url = "https://api.tosspayments.com/v1/brandpay/authorizations/access-token";
+        JSONObject response = sendRequest(requestData, API_SECRET_KEY, url);
+
+        logger.info("Response Data: {}", response);
+
+        return ResponseEntity.status(response.containsKey("error") ? 400 : 200).body(response);
+    }
+
+    @RequestMapping(value = "/confirm/brandpay", method = RequestMethod.POST, consumes = "application/json")
+    public ResponseEntity<JSONObject> confirmBrandpay(@RequestBody String jsonBody) throws Exception {
+        JSONObject requestData = parseRequestData(jsonBody);
+        String url = "https://api.tosspayments.com/v1/brandpay/payments/confirm";
+        JSONObject response = sendRequest(requestData, API_SECRET_KEY, url);
+        return ResponseEntity.status(response.containsKey("error") ? 400 : 200).body(response);
+    }
+
+    private JSONObject parseRequestData(String jsonBody) {
+        try {
+            return (JSONObject) new JSONParser().parse(jsonBody);
+        } catch (ParseException e) {
+            logger.error("JSON Parsing Error", e);
+            return new JSONObject();
+        }
+    }
+
+    private JSONObject sendRequest(JSONObject requestData, String secretKey, String urlString) throws IOException {
+        HttpURLConnection connection = createConnection(secretKey, urlString);
+        try (OutputStream os = connection.getOutputStream()) {
+            os.write(requestData.toString().getBytes(StandardCharsets.UTF_8));
+        }
+
+        try (InputStream responseStream = connection.getResponseCode() == 200 ? connection.getInputStream() : connection.getErrorStream();
+             Reader reader = new InputStreamReader(responseStream, StandardCharsets.UTF_8)) {
+            return (JSONObject) new JSONParser().parse(reader);
+        } catch (Exception e) {
+            logger.error("Error reading response", e);
+            JSONObject errorResponse = new JSONObject();
+            errorResponse.put("error", "Error reading response");
+            return errorResponse;
+        }
+    }
+
+    private HttpURLConnection createConnection(String secretKey, String urlString) throws IOException {
+        URL url = new URL(urlString);
+        HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+        connection.setRequestProperty("Authorization", "Basic " + Base64.getEncoder().encodeToString((secretKey + ":").getBytes(StandardCharsets.UTF_8)));
+        connection.setRequestProperty("Content-Type", "application/json");
+        connection.setRequestMethod("POST");
+        connection.setDoOutput(true);
+        return connection;
+    }
+
+    @RequestMapping(value = "/", method = RequestMethod.GET)
+    public String index() {
+        return "/widget/checkout";
+    }
+
+    @RequestMapping(value = "/fail", method = RequestMethod.GET)
+    public String failPayment(HttpServletRequest request, Model model) {
+        model.addAttribute("code", request.getParameter("code"));
+        model.addAttribute("message", request.getParameter("message"));
+        return "/fail";
+    }
+}
