+++ payment/src/main/java/com/carrot/payment/kafka/KafkaConsumer.java
@@ -1,6 +1,5 @@
 package com.carrot.payment.kafka;
 
-import com.carrot.payment.domain.Payment;
 import com.carrot.payment.service.PaymentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -19,21 +18,6 @@ public class KafkaConsumer {
 
         // 메시지의 action에 따라 적절한 PaymentService 메서드 호출
         switch (message.getAction()) {
-            case "create":
-                // create action에 대한 처리
-                Payment payment = message.getPayment();
-                service.startPayment(payment.getChatRoomId(), payment.getUserId(), payment.getPrice());
-                break;
-
-            case "approve":
-                // approve action에 대한 처리
-                service.approvePayment(message.getPayment().getId());
-                break;
-
-            case "cancel":
-                // cancel action에 대한 처리
-                service.cancelPayment(message.getPayment().getId());
-                break;
 
             default:
                 throw new IllegalArgumentException("Unknown action: " + message.getAction());
