+++ payment/src/main/java/com/carrot/payment/service/PaymentService.java
@@ -4,12 +4,6 @@ import com.carrot.payment.domain.Payment;
 import java.util.List;
 
 public interface PaymentService {
-    Payment startPayment(Long chatRoomId, Long userId, Long price);
-
-    void approvePayment(Long paymentId);
-
-    void cancelPayment(Long paymentId);
-
     List<Payment> getUserPaymentsHistory(Long userId);
 
     List<Payment> getPendingPayments();
